opencode/pyproject.toml
[project]
name = "opencode-workspace"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "jinja2",
    "jsonschema",
    "litellm",
    "pydantic",
    "pydantic-ai",
    "pyyaml",
    "rich",
    "workspace-scripts",
]

[dependency-groups]
dev = [
    "pytest",
]

[tool.uv.sources]
workspace-scripts = { path = "../scripts", editable = true }
